Revert "Remove generic type from `BaseApiClient`"
This change was intended to make building a client simpler, but it ended
up being more complicated than expected, using the async stub is more
complicated than expected, as it lives only in the `.pyi` file and can't
be used in any other context than type hints. For example the new
approach didn't worked well with delaying the connection of the client.
To handle that correctly, more work is needed by subclasses.

This commit reverts back to making the `BaseApiClient` class generic and
it instantiates the stub internally as before. To get proper async type
hints, users now only need to write the `stub` property themselves, and
use the appropriate async stub type hint there.

This reverts commit 035a7949a9bea7237911f64450a4200ba5d38eb7.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>