Change interface for HMAC
This changes the interface for communicating the API key and the signing
secret. Specifically it removes them from "authentication options" and
moves them to their own parameters instead.
This allows using the default pattern for channel options in places that
previously did not touch it. This is desired as soon most clients will
require the API key and the signing secret to function properly, and we
want to keep the changes needed to them minimal.

Signed-off-by: Florian Wagner <florian.wagner@frequenz.com>