Do a proper full comparison for events in tests
Only the types of the events were checked by the tests, not their state,
which doesn't really validate everything works properly.

We couldn't do a proper full comparison because we were using mocks
inside the error object, and mocks have non-deterministic
representations, as they have a unique ID.

To fix this we just create a full gRPC error instance. This is not
enough though, we also need to use a specific instance because 2 gRPC
error created the same are not equal.

Finally, we also need to set the retry strategy jitter to 0 so we can
make sure we have a predictable retry_time.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>