Replace `StreamStopped` with `StreamRetrying`
The new event is sent only when the stream will retry reconnecting
instead of every time the stream stops, as this information is less
important. When the stream stops, 2 things can happen, it is retried or
it finishes. The new event covers the retry case, and the case when
it finishes is covered by the receiver being closed.

With the current approach, we also needed to consume from the retry
strategy even if no retry was going to be attempted, so with this change
we can now only consume from the retry strategy when we actually want
to retry.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>