Remove wrong extra messages from test
Before the changes to the errored stream tests, the generator used for
error tests was not re-created each time the streamer reconnected. This
means that after the first error, the stream would resume correctly.

Before this was not an issue because the retry strategy was never reset,
so when retry was set to 2, it didn't matter that the stream was resumed
successfully, the next retry would still reduce the number or further
retry attempts.

Now that the retry strategy is properly reset, we need to make sure to
keep the errored channel in error state, otherwise the retry counter
will always be reset and never give up.

Because of this, now we don't receive messages anymore in the test after
the first failure.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>