Don't require using literals for `new_receiver()`
The overloads for `new_receiver()`'s `include_events` parameter used
`Literal[True]` and `Literal[False]` to differentiate the return types,
but this requires users to call the method only using literals, which is
a bit too restrictive.

This commit changes the type hints to use just `bool` for the parameter
for the general case. It also simplifies the type hints for the return
type.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>