Wait for the initial metadata before sending `StreamStarted` event
Sending the `StreamStarted` event as soon as the streaming method is
called can give a false impression that the stream is active and working,
when in reality it might not be connected to the server at all.

Now we wait until we receive the initial metadata from the server before
firing the `StreamStarted` event. That should give users a better
indication that the stream is actually active and working without having
to wait for the first item to be received, which can take a long time
for some low-frequency streams.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>